Add values to control the MaaS networking config and have the shell script work in conjunction with them. Keep in mind that if you change the defaults in values.yaml, you will also have to change the defaults in the script!

+++ build-cluster-application.sh
@@ -14,17 +14,26 @@ NEWFILE=cluster-applications/${BASE_URL:6}-$(basename $1)
 cp $1 $NEWFILE
 echo "New file location is $NEWFILE"
 
-SERVICE_CA_FILE=$(mktemp)
-oc get configmap signing-cabundle -n openshift-service-ca \
-  -o jsonpath='{.data.ca-bundle\.crt}' > "$SERVICE_CA_FILE"
-
 yq -i ".spec.source.helm.valuesObject.global.clusterBaseUrl = \"$BASE_URL\"" "$NEWFILE"
-# Literal block style preserves PEM newlines. Single-quoted multi-line YAML folds
-# them to spaces, which breaks OpenShift destinationCACertificate parsing.
-yq -i "
-  .spec.source.helm.valuesObject.global.serviceCABundle = load_str(\"$SERVICE_CA_FILE\") |
-  .spec.source.helm.valuesObject.global.serviceCABundle style=\"literal\"
-" "$NEWFILE"
-rm -f "$SERVICE_CA_FILE"
 
-oc apply -f $NEWFILE
+MAAS_ROUTE=$(yq '.spec.source.helm.valuesObject.configuration.inferenceApp.maasRoute // "false"' "$NEWFILE")
+MAAS_ROUTE_TERMINATION=$(yq '.spec.source.helm.valuesObject.configuration.inferenceApp.maasRouteConfig.termination // "passthrough"' "$NEWFILE")
+echo $MAAS_ROUTE
+echo $MAAS_ROUTE_TERMINATION
+if [[ $MAAS_ROUTE == "true" && $MAAS_ROUTE_TERMINATION == "reencrypt" ]]
+then
+  echo "Using routes with reencryption requires the cert to be set. Doing that now"
+  SERVICE_CA_FILE=$(mktemp)
+  oc get configmap signing-cabundle -n openshift-service-ca \
+    -o jsonpath='{.data.ca-bundle\.crt}' > "$SERVICE_CA_FILE"
+  
+  # Literal block style preserves PEM newlines. Single-quoted multi-line YAML folds
+  # them to spaces, which breaks OpenShift destinationCACertificate parsing.
+  yq -i "
+    .spec.source.helm.valuesObject.global.serviceCABundle = load_str(\"$SERVICE_CA_FILE\") |
+    .spec.source.helm.valuesObject.global.serviceCABundle style=\"literal\"
+  " "$NEWFILE"
+  rm -f "$SERVICE_CA_FILE"
+fi
+
+#oc apply -f $NEWFILE
